Rename: it works on people not just patients
